Mercury's Company Update
Version 1.1.0 the starting edition, this is the base of the mod for now with other things to come. I don't want to relay with Custom Sounds mod to keep my audio in game so will try to figure out the way to do other audio. Models are something i want to do next before i just do other mods and update this with other things i find funny.